# spring-security/documentation/swagger.yml
openapi: 3.0.0
info:
  version: '1.0'
  title: REST API Advanced
  description: Documentation REST API
servers:
  - url: http://localhost:8080
paths:
...
  /users:
    get:
      tags:
        - Users
      description: Returns all users
      operationId: getAllUsers
      summary: Returns all users
      parameters:
        - name: page
          in: query
          description: Page
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Size
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 20
      responses:
        '200':
          description: Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          links:
            GetUserById:
              operationId: getUserById
              parameters:
                id: 0
              description: Returns user by ID
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity404'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity500'
          description: Server Error
  /users/{id}:
    get:
      tags:
        - Users
      summary: Returns a user by ID.
      description: Returns user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          links:
            GetAllUsers:
              operationId: getAllUsers
              parameters:
                page: 0
                size: 20
              description: Returns all users
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity404'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity500'
          description: Server Error
  /orders/{userId}/{giftCertificateId}:
    post:
      tags:
        - Orders
      description: Returns created order
      summary: Returns created order
      operationId: createOrder
      parameters:
        - name: userId
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
        - name: giftCertificateId
          in: path
          description: Gift-certificate ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          links:
            GetOrderByOrderId:
              operationId: getOrderById
              parameters:
                orderId: 0
              description: Returns order by ID

            GetAllOrders:
              operationId: getAllOrders
              parameters:
                page: 0
                size: 20
              description: Returns all Orders

        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity404'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity500'
          description: Server Error
  /orders:
    get:
      tags:
        - Orders
      description: Returns order
      summary: Returns order
      operationId: getAllOrders
      parameters:
        - name: page
          in: query
          description: Page
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Size
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 20
      responses:
        '200':
          description: Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
                # remake
          links:
            GetOrderByOrderId:
              operationId: getOrderById
              parameters:
                orderId: 0
              description: Returns order by ID

            CreateOrder:
              operationId: createOrder
              parameters:
                userId: 0
                giftCertificateId: 0
              description: Returns created order

        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity404'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity500'
          description: Server Error
  /orders/{id}:
    get:
      tags:
        - Orders
      description: Returns order by ID
      summary: Returns order by ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

          links:
            CreateOrder:
              operationId: createOrder
              parameters:
                userId: 0
                giftCertificateId: 0
              description: Returns created order

            GetAllOrders:
              operationId: getAllOrders
              parameters:
                page: 0
                size: 20
              description: Returns all orders

          description: Status OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity404'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEntity500'
          description: Server Error
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
      required:
        - id
        - name
    Gift-certificate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: integer
          format: int64
        durationDays:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        createDate:
          type: string
          format: date-time
        lastUpdateDate:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - durationDays
        - createDate
        - lastUpdateDate
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
        - id
        - name
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cost:
          type: integer
          format: int128
        gift-certificate:
          type: array
          description: Order`s gift-certificate
          items:
            $ref: '#/components/schemas/Gift-certificate'
        purchaseDate:
          type: string
          format: date-time
      required:
        - id
        - cost
        - gift-certificate
        - purchaseDate
    ResponseEntity404:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - 404 NOT_FOUND
        response body:
          type: array
          items:
            $ref: '#/components/schemas/messageResponseNotFound'
    messageResponseNotFound:
      type: object
      properties:
        message:
          type: string
          example: "NOT_FOUND"
    ResponseEntity500:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - 500 INTERNAL_SERVER_ERROR
        response body:
          type: array
          items:
            $ref: '#/components/schemas/messageResponseServerError'
    messageResponseServerError:
      type: object
      properties:
        message:
          type: string
          example: "INTERNAL_SERVER_ERROR"